requirements.txt
python-telegram-bot[job-queue]==20.7
playwright==1.48.0
